Remove num argument from subSqes and add return value
This make it possible to even detect if io_uring_enter fails.
`num` was removed, because I don't think you would ever want to submit
less, than you can. Maybe with links? Also I keep track of the flushed
elements and the submitted elements. Previously if you flushed an
element that for some reason could not be submitted, you would be
"behind" forever. Now those elements would be submitted (or tried to)
the next time submitSqes is called.